Add list, recreate project because of crash

+++ assets/db/init_script.sql
@@ -1,11 +1,11 @@
 CREATE TABLE "Expenses" (
-	"id"	INTEGER NOT NULL,
-	"cost"	INTEGER NOT NULL,
+	"id"	    INTEGER NOT NULL,
+	"cost"	    REAL NOT NULL,
 	"dateTime"	TEXT NOT NULL,
-	"desc"	TEXT,
-	"name"	TEXT,
-	"tag"	INTEGER DEFAULT 0,
-	"owner"	INTEGER DEFAULT 0,
+	"desc"	    TEXT,
+	"name"	    TEXT,
+	"tag"	    INTEGER DEFAULT 0,
+	"owner"	    INTEGER DEFAULT 0,
 	PRIMARY KEY("id" AUTOINCREMENT),
 	FOREIGN KEY("owner") REFERENCES "owners"("id") ON UPDATE SET DEFAULT ON DELETE SET DEFAULT,
 	FOREIGN KEY("tag") REFERENCES "tags"("id") ON UPDATE SET DEFAULT ON DELETE SET DEFAULT
@@ -18,7 +18,7 @@ CREATE TABLE "Owners" (
 );
 
 INSERT INTO "Owners" ("id", "name")
-VALUES (0, "no_name");
+VALUES (0, 'no_name');
 
 CREATE TABLE "Tags" (
 	"id"	INTEGER NOT NULL,
@@ -27,4 +27,4 @@ CREATE TABLE "Tags" (
 );
 
 INSERT INTO "Tags" ("id", "name")
-VALUES (0, "");
+VALUES (0, '');
